requirements.txt
Flask
flask-cors
docx2pdf
PyPDF2
docxcompose
tabula-py
pandas
openpyxl
gunicorn
python-docx